+++ src/main/java/lagatrix/server/entities/connection/Request.java
@@ -1,5 +1,6 @@
 package lagatrix.server.entities.connection;
 
+import java.util.Arrays;
 import lagatrix.server.entities.actions.ActionsEnum;
 
 /**
@@ -11,10 +12,19 @@ import lagatrix.server.entities.actions.ActionsEnum;
 public class Request {
     
     private ActionsEnum action;
+    private Class objectWhoRequest;
     private Object[] params;
 
-    public Request(ActionsEnum action, Object... params) {
+    /**
+     * The constructor of the class.
+     * 
+     * @param action The action to exec.
+     * @param objectWhoRequest The item who manage.
+     * @param params If the execuniton need params.
+     */
+    public Request(ActionsEnum action, Class objectWhoRequest, Object... params) {
         this.action = action;
+        this.objectWhoRequest = objectWhoRequest;
         this.params = params;
     }
     
@@ -27,7 +37,17 @@ public class Request {
         return action;
     }
 
+    public Class getObjectWhoRequest() {
+        return objectWhoRequest;
+    }
+    
     public Object[] getParams() {
         return params;
+    }
+
+    @Override
+    public String toString() {
+        return "Request{" + "action=" + action + ", objectWhoRequest=" + objectWhoRequest.getSimpleName()
+                + ", params=" + Arrays.toString(params) + '}';
     }
 }